feat: update contributor class
The price, fee and startTime got removed from the contributor class.
These all got moved to a global state. Next to this, a name and
cliff got added to the contributor class. All functions related to
modifiying the classes got updated as well.

Because the fee (soon tipping) will be moved to a global tipping
structure that is sperate of the funders, I removed all the fee
mentioning in the functions such as buy() and claim(), as well as
the functions to calculate the fee from the funders