feat: update class creation and allowlisting
This change was made because the whitelist mapping was seperate
from the FundersPortfolio. Because of that someone could be part
of a class but not whitelisted etc. To make this easier the
whitelist (allowlist) was moved to the funders struct. Inside the
function is checked if allowlisting need to happen.

Next to this, a function got added that will be used to add classes
at the same time with allowlists. Only this function will be called
so the same function gets called for adding 1 or 60 classes