feat: seed tip structure
The protocol fee is now considerd a tip because it is voluntary given.
The tip is now no longer part of the buyable Seed amount but a
sperate amount, which can never be bought or claimend back by
the admin in case of a failed Seed i.e softCap not reached. The
tip has it's own vesting cliff & duration and can be claimed through
a seperate function