fix: update fundingAmount in buy function dynamically
If the fundingAmount was to high when buying, it would revert the
tx in the past. This could be exploited for a griefing attack
whereby a tx could be frontrun, and always revert because of added
funding amount. This could lead to a Seed never reachding hardCap.

The change made will automatically adapt the funding amount if
contributing the amount would exeed hardCap