chore: set vestingStartTime when Seed is closed
**Why**
The beneficiary and buyers (if applicable) would need to wait
until endTime is reached before vesting starts, in case the admin
closes the Seed. This is bad UX.
**How**
Changing the code so the vestingStartTime is set to the current
time when the Seed is closed by the Admin.
**What**
In the function `close()`, set the vestingStartTime if it is smaller
then the current time, i.e. it is still set to the endTime and
endTime has not been reached.

**Note**
Also moved require statements to a modifier to clean the code