fix: restrict adding/editing classes to be able until startTime
- Why
To prohibit a malicious admin from changing class parameters after
a buyer has contributed token. Example could be setting the
cliff to something like max int. Changing and adding a class can
now only happen until startTime
- How
Restricting the option to add and or edit a class after startTime
and when a Seed is closed.
- What
Add modifiers to check if the Seed is closed and check that
startTime has not been reached