feat: add function to change class and allowlist
- Why
To simplify the UX because right now changing a class and adding
an allowlist are two contract calls, but for the FE this is one
step.
- How
Adding a new function which replaces the old changeClass() function
- What
Renamed function and added a allowlist parameter to it. Add conditional
adding of addresses if allowlist is not empty.