feat: remove access control from function deploySeed()
- Why
Because the Celo Safe is flaky, we cannot rely on it behaving as
expected. We adapt the contract in a way that will remove the Safe
from our deployment flow.
- How
To resolve this dependency we have removed the access control from
the function deploySeed(). This means that everyone can directly deploy
a Seed from through the FE for now.
- What
Remove the modifier in the contract, renamed the file and added a
comment in the file explaining what has been done.